backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-unfold==0.40.0
python-dotenv==1.0.1
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
black==24.2.0
